@@ -14,8 +14,8 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "twig/twig": "^2.11",
+        "php": ">=7.2.5",
+        "twig/twig": "~3.0",
         "illuminate/support": "^5.5|^6|^7|^8",
         "illuminate/view": "^5.5|^6|^7|^8"
     },
