@@ -44,10 +44,6 @@ return [
             // default: utf-8
             'charset' => 'utf-8',
 
-            // The base template class to use for generated templates.
-            // default: TwigBridge\Twig\Template
-            'base_template_class' => 'TwigBridge\Twig\Template',
-
             // An absolute path where to store the compiled templates, or false to disable caching. If null
             // then the cache file path is used.
             // default: cache file storage path
@@ -111,9 +107,11 @@ return [
         |
         */
         'enabled' => [
+            'TwigBridge\Extension\Laravel\Event',
             'TwigBridge\Extension\Loader\Facades',
             'TwigBridge\Extension\Loader\Filters',
             'TwigBridge\Extension\Loader\Functions',
+            'TwigBridge\Extension\Loader\Globals',
 
             'TwigBridge\Extension\Laravel\Auth',
             'TwigBridge\Extension\Laravel\Config',
