@@ -12,11 +12,11 @@
 namespace TwigBridge;
 
 use Illuminate\Contracts\Container\Container;
-use Illuminate\View\ViewFinderInterface;
 use InvalidArgumentException;
 use Twig\Environment;
 use Twig\Error\Error;
 use Twig\Loader\LoaderInterface;
+use Twig\Source;
 
 /**
  * Bridge functions between Laravel & Twig
@@ -70,15 +70,6 @@ class Bridge extends Environment
         $this->app = $app;
     }
 
-    public function loadTemplate($name, $index = null)
-    {
-        $template = parent::loadTemplate($name, $index);
-
-        $template->setName($this->normalizeName($name));
-
-        return $template;
-    }
-
     /**
      * Lint (check) the syntax of a file on the view paths.
      *
@@ -88,10 +79,12 @@ class Bridge extends Environment
      */
     public function lint($file)
     {
+        /** @var Source $template */
         $template = $this->app['twig.loader.viewfinder']->getSourceContext($file);
 
-        if (!$template) {
-            throw new InvalidArgumentException('Unable to find file: '.$file);
+        $code = trim($template->getCode());
+        if (empty($code)) {
+            throw new InvalidArgumentException('Unable to find file: ' . $file);
         }
 
         try {
@@ -101,52 +94,5 @@ class Bridge extends Environment
         }
 
         return true;
-    }
-
-    /**
-     * Merges a context with the shared variables, same as mergeGlobals()
-     *
-     * @param array $context An array representing the context
-     *
-     * @return array The context merged with the globals
-     */
-    public function mergeShared(array $context)
-    {
-        // we don't use array_merge as the context being generally
-        // bigger than globals, this code is faster.
-        foreach ($this->app['view']->getShared() as $key => $value) {
-            if (!array_key_exists($key, $context)) {
-                $context[$key] = $value;
-            }
-        }
-
-        return $context;
-    }
-
-    /**
-     * Normalize a view name.
-     *
-     * @param  string $name
-     *
-     * @return string
-     */
-    protected function normalizeName($name)
-    {
-        $extension = '.' . $this->app['twig.extension'];
-        $length = strlen($extension);
-
-        if (substr($name, -$length, $length) === $extension) {
-            $name = substr($name, 0, -$length);
-        }
-
-        // Normalize namespace and delimiters
-        $delimiter = ViewFinderInterface::HINT_PATH_DELIMITER;
-        if (strpos($name, $delimiter) === false) {
-            return str_replace('/', '.', $name);
-        }
-
-        list($namespace, $name) = explode($delimiter, $name);
-
-        return $namespace.$delimiter.str_replace('/', '.', $name);
     }
 }
