@@ -4,7 +4,8 @@ namespace TwigBridge\Tests\Command\Lint;
 
 use Mockery as m;
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Output\StreamOutput;
+use Twig\Error\LoaderError;
+use Twig\Loader\LoaderInterface;
 use TwigBridge\Command\Lint;
 
 class ContentTest extends Base
@@ -24,8 +25,8 @@ class ContentTest extends Base
         $command = new Lint;
         $app     = $this->getApplication();
 
-        $loader = m::mock('Twig_LoaderInterface');
-        $loader->shouldReceive('getSourceContext')->andThrow(new \Twig_Error_Loader('test'));
+        $loader = m::mock(LoaderInterface::class);
+        $loader->shouldReceive('getSourceContext')->andThrow(new LoaderError('test'));
         $app['twig.loader'] = $loader;
 
         $command->setLaravel($app);
