@@ -2,9 +2,10 @@
 
 namespace TwigBridge\Tests\ServiceProvider\Bindings;
 
-use TwigBridge\Tests\Base;
 use Mockery as m;
+use Twig\Lexer;
 use TwigBridge\ServiceProvider;
+use TwigBridge\Tests\Base;
 
 class BridgeTest extends Base
 {
@@ -23,7 +24,7 @@ class BridgeTest extends Base
     public function testSetLexer()
     {
         $app = $this->getApplication();
-        $app['twig.lexer'] = m::mock('Twig_Lexer');
+        $app['twig.lexer'] = m::mock(Lexer::class);
         $app['twig.lexer']->shouldReceive('fooBar')->andReturn('baz');
         $app['twig.extensions'] = [];
 
